+++ Tests/Get-PASOpenIDConnectProvider.Tests.ps1
@@ -36,12 +36,18 @@ Describe $($PSCommandPath -Replace ".Tests.ps1") {
 	InModuleScope $(Split-Path (Split-Path (Split-Path -Parent $PSCommandPath) -Parent) -Leaf ) {
 
 		BeforeEach {
+			$Script:ExternalVersion = "0.0"
+
 			Mock Invoke-PASRestMethod -MockWith {
-				[PSCustomObject]@{"Methods" = [PSCustomObject]@{"Prop1" = "Val1"; "Prop2" = "Val2" } }
+
+				[PSCustomObject]@{"Providers" = [PSCustomObject]@{"Prop1" = "Val1"; "Prop2" = "Val2" } }
+
 			}
 
-			$response = Get-PASAuthenticationMethod
+			$response = Get-PASOpenIDConnectProvider
+
 		}
+
 		Context "Input" {
 
 			It "sends request" {
@@ -54,7 +60,7 @@ Describe $($PSCommandPath -Replace ".Tests.ps1") {
 
 				Assert-MockCalled Invoke-PASRestMethod -ParameterFilter {
 
-					$URI -eq "$($Script:BaseURI)/api/Configuration/AuthenticationMethods/"
+					$URI -eq "$($Script:BaseURI)/api/Configuration/OIDC/Providers/"
 
 				} -Times 1 -Exactly -Scope It
 
@@ -62,11 +68,11 @@ Describe $($PSCommandPath -Replace ".Tests.ps1") {
 
 			It "sends request to expected endpoint - get specific method" {
 
-				Get-PASAuthenticationMethod -ID SomeMethod
+				Get-PASOpenIDConnectProvider -id idValue
 
 				Assert-MockCalled Invoke-PASRestMethod -ParameterFilter {
 
-					$URI -eq "$($Script:BaseURI)/api/Configuration/AuthenticationMethods/SomeMethod"
+					$URI -eq "$($Script:BaseURI)/api/Configuration/OIDC/Providers/idValue"
 
 				} -Times 1 -Exactly -Scope It
 
@@ -86,7 +92,7 @@ Describe $($PSCommandPath -Replace ".Tests.ps1") {
 
 			It "throws error if version requirement not met" {
 				$Script:ExternalVersion = "1.0"
-				{ Get-PASAuthenticationMethod } | Should -Throw
+				{ $InputObject | Get-PASOpenIDConnectProvider } | Should -Throw
 				$Script:ExternalVersion = "0.0"
 			}